requirements.txt
# Stock data collection
yfinance>=0.2.36

# News and web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
feedparser>=6.0.10

# Korean DART (공시) data
dart-fss==0.4.15

# Data manipulation and analysis
pandas>=2.2.0
numpy>=1.26.0

# AI API
anthropic>=0.40.0

# Web Dashboard
streamlit>=1.31.0
plotly>=5.18.0
kaleido>=0.2.1

# PDF & Email
fpdf2>=2.7.0
Pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0

# Optional but useful
openpyxl>=3.1.0
python-dateutil>=2.8.0
